backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.8
pydantic==2.5.3
